feat: Integrate BenchmarkingProtocol into ExperimentProtocol
