+++ src/qubex/experiment/mixin/__init__.py
@@ -1,8 +1,10 @@
+from .benchmarking_mixin import BenchmarkingMixin
 from .calibration_mixin import CalibrationMixin
 from .characterization_mixin import CharacterizationMixin
 from .measurement_mixin import MeasurementMixin
 
 __all__ = [
+    "BenchmarkingMixin",
     "CalibrationMixin",
     "CharacterizationMixin",
     "MeasurementMixin",
