+++ src/qubex/experiment/mixin/calibration_mixin.py
@@ -1,42 +1,27 @@
 from __future__ import annotations
 
 from collections import defaultdict
-from typing import Collection, Literal, Sequence
+from typing import Collection, Literal
 
 import cma
 import numpy as np
 import plotly.graph_objects as go
-from numpy.typing import ArrayLike, NDArray
-from tqdm import tqdm
+from numpy.typing import ArrayLike
 
 from ...analysis import fitting
 from ...analysis import visualization as vis
 from ...backend import SAMPLING_PERIOD, Target
-from ...clifford import Clifford
-from ...measurement import (
-    MeasureResult,
-    StateClassifier,
-    StateClassifierGMM,
-    StateClassifierKMeans,
-)
 from ...measurement.measurement import DEFAULT_INTERVAL, DEFAULT_SHOTS
 from ...pulse import (
     CrossResonance,
     Drag,
     FlatTop,
-    PhaseShift,
     Pulse,
     PulseSchedule,
     PulseSequence,
-    VirtualZ,
     Waveform,
 )
-from ...typing import (
-    IQArray,
-    ParametricPulseSchedule,
-    ParametricWaveformDict,
-    TargetMap,
-)
+from ...typing import TargetMap
 from ..experiment_constants import (
     CALIBRATION_SHOTS,
     CR_PARAMS,
@@ -53,9 +38,8 @@ from ..experiment_constants import (
     PI_AMPLITUDE,
     PI_DURATION,
     PI_RAMPTIME,
-    STATE_CENTERS,
 )
-from ..experiment_result import AmplCalibData, ExperimentResult, RBData
+from ..experiment_result import AmplCalibData, ExperimentResult
 from ..protocol import BaseProtocol, CalibrationProtocol, MeasurementProtocol
 
 
@@ -73,29 +57,6 @@ class CalibrationMixin(
         shots: int = DEFAULT_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> ExperimentResult[AmplCalibData]:
-        """
-        Calibrates the default pulse.
-
-        Parameters
-        ----------
-        targets : Collection[str]
-            Target qubits to calibrate.
-        pulse_type : Literal["pi", "hpi"]
-            Type of the pulse to calibrate.
-        n_rotations : int, optional
-            Number of rotations. Defaults to 1.
-        plot : bool, optional
-            Whether to plot the measured signals. Defaults to True.
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        ExperimentResult[AmplCalibData]
-            Result of the experiment.
-        """
         targets = list(targets)
         rabi_params = self.rabi_params
         if rabi_params is None:
@@ -170,27 +131,6 @@ class CalibrationMixin(
         shots: int = DEFAULT_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> ExperimentResult[AmplCalibData]:
-        """
-        Calibrates the default pulse.
-
-        Parameters
-        ----------
-        targets : Collection[str]
-            Target qubits to calibrate.
-        pulse_type : Literal["pi", "hpi"]
-            Type of the pulse to calibrate.
-        n_rotations : int, optional
-            Number of rotations. Defaults to 1.
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        ExperimentResult[AmplCalibData]
-            Result of the experiment.
-        """
         targets = list(targets)
         rabi_params = self.rabi_params
         if rabi_params is None:
@@ -285,37 +225,6 @@ class CalibrationMixin(
         shots: int = CALIBRATION_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> dict[str, float]:
-        """
-        Calibrates the DRAG amplitude.
-
-        Parameters
-        ----------
-        targets : Collection[str]
-            Target qubits to calibrate.
-        spectator_state : str, optional
-            Spectator state. Defaults to "+".
-        pulse_type : Literal["pi", "hpi"]
-            Type of the pulse to calibrate.
-        n_rotations : int, optional
-            Number of rotations to |0> state. Defaults to 4.
-        duration : float, optional
-            Duration of the pulse. Defaults to None.
-        drag_coeff : float, optional
-            DRAG coefficient. Defaults to DRAG_COEFF.
-        use_stored_amplitude : bool, optional
-            Whether to use the stored amplitude. Defaults to False.
-        use_stored_beta : bool, optional
-            Whether to use the stored beta. Defaults to False.
-        shots : int, optional
-            Number of shots. Defaults to CALIBRATION_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        dict[str, float]
-            Result of the calibration.
-        """
         targets = list(targets)
         rabi_params = self.rabi_params
         self.validate_rabi_params(rabi_params)
@@ -438,35 +347,6 @@ class CalibrationMixin(
         shots: int = CALIBRATION_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> dict[str, float]:
-        """
-        Calibrates the DRAG beta.
-
-        Parameters
-        ----------
-        targets : Collection[str]
-            Target qubits to calibrate.
-        spectator_state : str, optional
-            Spectator state. Defaults to "+".
-        pulse_type : Literal["pi", "hpi"]
-            Type of the pulse to calibrate.
-        beta_range : ArrayLike, optional
-            Range of the beta to sweep. Defaults to np.linspace(-0.5, 1.5, 41).
-        n_turns : int, optional
-            Number of turns to |0> state. Defaults to 1.
-        duration : float, optional
-            Duration of the pulse. Defaults to None.
-        degree : int, optional
-            Degree of the polynomial to fit. Defaults to 3.
-        shots : int, optional
-            Number of shots. Defaults to CALIBRATION_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        dict[str, float]
-            Result of the calibration.
-        """
         if targets is None:
             targets = self.qubit_labels
         else:
@@ -573,25 +453,6 @@ class CalibrationMixin(
         shots: int = CALIBRATION_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> ExperimentResult[AmplCalibData]:
-        """
-        Calibrates the π/2 pulse.
-
-        Parameters
-        ----------
-        targets : Collection[str], optional
-            Target qubits to calibrate.
-        n_rotations : int, optional
-            Number of rotations. Defaults to 1.
-        shots : int, optional
-            Number of shots. Defaults to CALIBRATION_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        ExperimentResult[AmplCalibData]
-            Result of the experiment.
-        """
         if targets is None:
             targets = self.qubit_labels
         else:
@@ -617,25 +478,6 @@ class CalibrationMixin(
         shots: int = CALIBRATION_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> ExperimentResult[AmplCalibData]:
-        """
-        Calibrates the π pulse.
-
-        Parameters
-        ----------
-        targes : Collection[str], optional
-            Target qubits to calibrate.
-        n_rotations : int, optional
-            Number of rotations. Defaults to 1.
-        shots : int, optional
-            Number of shots. Defaults to CALIBRATION_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        ExperimentResult[AmplCalibData]
-            Result of the experiment.
-        """
         if targets is None:
             targets = self.qubit_labels
         else:
@@ -661,25 +503,6 @@ class CalibrationMixin(
         shots: int = CALIBRATION_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> ExperimentResult[AmplCalibData]:
-        """
-        Calibrates the ef π/2 pulse.
-
-        Parameters
-        ----------
-        targets : Collection[str], optional
-            Target qubits to calibrate.
-        n_rotations : int, optional
-            Number of rotations. Defaults to 1.
-        shots : int, optional
-            Number of shots. Defaults to CALIBRATION_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        ExperimentResult[AmplCalibData]
-            Result of the experiment.
-        """
         if targets is None:
             targets = self.qubit_labels
         else:
@@ -705,25 +528,6 @@ class CalibrationMixin(
         shots: int = CALIBRATION_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> ExperimentResult[AmplCalibData]:
-        """
-        Calibrates the ef π pulse.
-
-        Parameters
-        ----------
-        targets : Collection[str], optional
-            Target qubits to calibrate.
-        n_rotations : int, optional
-            Number of rotations. Defaults to 1.
-        shots : int, optional
-            Number of shots. Defaults to CALIBRATION_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        ExperimentResult[AmplCalibData]
-            Result of the experiment.
-        """
         if targets is None:
             targets = self.qubit_labels
         else:
@@ -755,37 +559,6 @@ class CalibrationMixin(
         shots: int = CALIBRATION_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> dict:
-        """
-        Calibrates the DRAG π/2 pulse.
-
-        Parameters
-        ----------
-        targets : Collection[str], optional
-            Target qubits to calibrate.
-        n_rotations : int, optional
-            Number of rotations to |0> state. Defaults to 4.
-        n_turns : int, optional
-            Number of turns to |0> state. Defaults to 1.
-        n_iterations : int, optional
-            Number of iterations. Defaults to 2.
-        calibrate_beta : bool, optional
-            Whether to calibrate the DRAG beta. Defaults to True.
-        beta_range : ArrayLike, optional
-            Range of the beta to sweep. Defaults to np.linspace(-0.5, 1.5, 41).
-        duration : float, optional
-            Duration of the pulse. Defaults to None.
-        drag_coeff : float, optional
-            DRAG coefficient. Defaults to DRAG_COEFF.
-        shots : int, optional
-            Number of shots. Defaults to CALIBRATION_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        dict
-            Result of the calibration.
-        """
         if targets is None:
             targets = self.qubit_labels
         else:
@@ -850,39 +623,6 @@ class CalibrationMixin(
         shots: int = CALIBRATION_SHOTS,
         interval: int = DEFAULT_INTERVAL,
     ) -> dict:
-        """
-        Calibrates the DRAG π pulse.
-
-        Parameters
-        ----------
-        targets : Collection[str], optional
-            Target qubits to calibrate.
-        n_rotations : int, optional
-            Number of rotations to |0> state. Defaults to 4.
-        n_turns : int, optional
-            Number of turns to |0> state. Defaults to 1.
-        n_iterations : int, optional
-            Number of iterations. Defaults to 2.
-        calibrate_beta : bool, optional
-            Whether to calibrate the DRAG beta. Defaults to False.
-        beta_range : ArrayLike, optional
-            Range of the beta to sweep. Defaults to np.linspace(-0.5, 1.5, 41).
-        duration : float, optional
-            Duration of the pulse. Defaults to None.
-        drag_coeff : float, optional
-            DRAG coefficient. Defaults to DRAG_COEFF.
-        degree : int, optional
-            Degree of the polynomial to fit. Defaults to 3.
-        shots : int, optional
-            Number of shots. Defaults to CALIBRATION_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        dict
-            Result of the calibration.
-        """
         if targets is None:
             targets = self.qubit_labels
         else:
@@ -932,1495 +672,6 @@ class CalibrationMixin(
             "amplitude": amplitude,
             "beta": beta,
         }
-
-    def measure_state_distribution(
-        self,
-        targets: Collection[str] | None = None,
-        *,
-        n_states: Literal[2, 3] = 2,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-    ) -> list[MeasureResult]:
-        if targets is None:
-            targets = self.qubit_labels
-        else:
-            targets = list(targets)
-
-        states = ["g", "e", "f"][:n_states]
-        result = {
-            state: self.measure_state(
-                {target: state for target in targets},  # type: ignore
-                shots=shots,
-                interval=interval,
-            )
-            for state in states
-        }
-        for target in targets:
-            data = {
-                f"|{state}⟩": result[state].data[target].kerneled for state in states
-            }
-            if plot:
-                vis.plot_state_distribution(
-                    data=data,
-                    title=f"State distribution : {target}",
-                )
-        return list(result.values())
-
-    def build_classifier(
-        self,
-        targets: Collection[str] | None = None,
-        *,
-        n_states: Literal[2, 3] = 2,
-        shots: int = 10000,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-    ) -> dict:
-        if targets is None:
-            targets = self.qubit_labels
-        else:
-            targets = list(targets)
-
-        results = self.measure_state_distribution(
-            targets=targets,
-            n_states=n_states,
-            shots=shots,
-            interval=interval,
-            plot=False,
-        )
-
-        data = {
-            target: {
-                state: result.data[target].kerneled
-                for state, result in enumerate(results)
-            }
-            for target in targets
-        }
-
-        classifiers: TargetMap[StateClassifier]
-        if self.classifier_type == "kmeans":
-            classifiers = {
-                target: StateClassifierKMeans.fit(data[target]) for target in targets
-            }
-        elif self.classifier_type == "gmm":
-            classifiers = {
-                target: StateClassifierGMM.fit(data[target]) for target in targets
-            }
-        else:
-            raise ValueError("Invalid classifier type.")
-        self.measurement.classifiers = classifiers
-
-        fidelities = {}
-        average_fidelities = {}
-        for target in targets:
-            clf = classifiers[target]
-            classified = []
-            for state in range(n_states):
-                if plot:
-                    print(f"{target} prepared as |{state}⟩:")
-                result = clf.classify(
-                    target,
-                    data[target][state],
-                    plot=plot,
-                )
-                classified.append(result)
-            fidelity = [
-                classified[state][state] / sum(classified[state].values())
-                for state in range(n_states)
-            ]
-            fidelities[target] = fidelity
-            average_fidelities[target] = np.mean(fidelity)
-
-            if plot:
-                print(f"{target}:")
-                print(f"  Total shots: {shots}")
-                for state in range(n_states):
-                    print(
-                        f"  |{state}⟩ → {classified[state]}, f_{state}: {fidelity[state] * 100:.2f}%"
-                    )
-                print(
-                    f"  Average readout fidelity : {average_fidelities[target] * 100:.2f}%\n\n"
-                )
-
-        self.system_note.put(
-            STATE_CENTERS,
-            {
-                target: {
-                    str(state): (center.real, center.imag)
-                    for state, center in classifiers[target].centers.items()
-                }
-                for target in targets
-            },
-        )
-
-        return {
-            "readout_fidelties": fidelities,
-            "average_readout_fidelity": average_fidelities,
-            "measure_results": results,
-            "classifiers": classifiers,
-        }
-
-    def rb_sequence(
-        self,
-        *,
-        target: str,
-        n: int,
-        x90: dict[str, Waveform] | None = None,
-        zx90: PulseSchedule | dict[str, Waveform] | None = None,
-        interleaved_waveform: (
-            PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform] | None
-        ) = None,
-        interleaved_clifford: Clifford | None = None,
-        seed: int | None = None,
-    ) -> PulseSchedule:
-        target_object = self.experiment_system.get_target(target)
-        if target_object.is_cr:
-            sched = self.rb_sequence_2q(
-                target=target,
-                n=n,
-                x90=x90,
-                zx90=zx90,
-                interleaved_waveform=interleaved_waveform,
-                interleaved_clifford=interleaved_clifford,
-                seed=seed,
-            )
-            return sched
-        else:
-            if isinstance(interleaved_waveform, PulseSchedule):
-                interleaved_waveform = interleaved_waveform.get_sequences()
-            seq = self.rb_sequence_1q(
-                target=target,
-                n=n,
-                x90=x90,
-                interleaved_waveform=interleaved_waveform,
-                interleaved_clifford=interleaved_clifford,
-                seed=seed,
-            )
-            with PulseSchedule([target]) as ps:
-                ps.add(target, seq)
-            return ps
-
-    def rb_sequence_1q(
-        self,
-        *,
-        target: str,
-        n: int,
-        x90: Waveform | dict[str, Waveform] | None = None,
-        interleaved_waveform: (
-            Waveform | dict[str, PulseSequence] | dict[str, Waveform] | None
-        ) = None,
-        interleaved_clifford: Clifford | dict[str, tuple[complex, str]] | None = None,
-        seed: int | None = None,
-    ) -> PulseSequence:
-        """
-        Generates a randomized benchmarking sequence.
-
-        Parameters
-        ----------
-        target : str
-            Target qubit.
-        n : int
-            Number of Clifford gates.
-        x90 : Waveform | dict[str, Waveform], optional
-            π/2 pulse used for the experiment. Defaults to None.
-        interleaved_waveform : Waveform | dict[str, PulseSequence] | dict[str, Waveform], optional
-            Waveform of the interleaved gate. Defaults to None.
-        interleaved_clifford : Clifford | dict[str, tuple[complex, str]], optional
-            Clifford map of the interleaved gate. Defaults to None.
-        seed : int, optional
-            Random seed.
-
-        Returns
-        -------
-        PulseSequence
-            Randomized benchmarking sequence.
-
-        Examples
-        --------
-        >>> sequence = ex.rb_sequence(
-        ...     target="Q00",
-        ...     n=100,
-        ...     x90=Rect(duration=30, amplitude=0.1),
-        ... )
-
-        >>> sequence = ex.rb_sequence(
-        ...     target="Q00",
-        ...     n=100,
-        ...     x90=Rect(duration=30, amplitude=0.1),
-        ...     interleaved_waveform=Rect(duration=30, amplitude=0.1),
-        ...     interleaved_clifford={
-        ...         "I": (1, "I"),
-        ...         "X": (1, "X"),
-        ...         "Y": (-1, "Y"),
-        ...         "Z": (-1, "Z"),
-        ...     },
-        ... )
-        """
-        if isinstance(x90, dict):
-            x90 = x90.get(target)
-        x90 = x90 or self.hpi_pulse[target]
-        z90 = VirtualZ(np.pi / 2)
-
-        sequence: list[Waveform | VirtualZ] = []
-
-        if interleaved_waveform is None:
-            cliffords, inverse = self.clifford_generator.create_rb_sequences(
-                n=n,
-                type="1Q",
-                seed=seed,
-            )
-        else:
-            if interleaved_clifford is None:
-                raise ValueError("`interleaved_clifford` must be provided.")
-            cliffords, inverse = self.clifford_generator.create_irb_sequences(
-                n=n,
-                interleave=interleaved_clifford,
-                type="1Q",
-                seed=seed,
-            )
-
-        def add_gate(gate: str):
-            if gate == "X90":
-                sequence.append(x90)
-            elif gate == "Z90":
-                sequence.append(z90)
-            else:
-                raise ValueError("Invalid gate.")
-
-        for clifford in cliffords:
-            for gate in clifford:
-                add_gate(gate)
-            if isinstance(interleaved_waveform, dict):
-                interleaved_waveform = interleaved_waveform.get(target)
-            if interleaved_waveform is not None:
-                sequence.append(interleaved_waveform)
-
-        for gate in inverse:
-            add_gate(gate)
-
-        return PulseSequence(sequence)
-
-    def rb_sequence_2q(
-        self,
-        *,
-        target: str,
-        n: int,
-        x90: TargetMap[Waveform] | None = None,
-        zx90: (
-            PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform] | None
-        ) = None,
-        interleaved_waveform: (
-            PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform] | None
-        ) = None,
-        interleaved_clifford: Clifford | dict[str, tuple[complex, str]] | None = None,
-        seed: int | None = None,
-    ) -> PulseSchedule:
-        """
-        Generates a 2Q randomized benchmarking sequence.
-
-        Parameters
-        ----------
-        target : str
-            Target qubit.
-        n : int
-            Number of Clifford gates.
-        x90 : Waveform | TargetMap[Waveform], optional
-            π/2 pulse used for 1Q gates. Defaults to None.
-        zx90 : PulseSchedule | dict[str, Waveform], optional
-            ZX90 pulses used for 2Q gates. Defaults to None.
-        interleaved_waveform : PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform], optional
-            Waveform of the interleaved gate. Defaults to None.
-        interleaved_clifford : Clifford | dict[str, tuple[complex, str]], optional
-            Clifford map of the interleaved gate. Defaults to None.
-        seed : int, optional
-            Random seed.
-
-        Returns
-        -------
-        PulseSchedule
-            Randomized benchmarking sequence.
-
-        Examples
-        --------
-        >>> sequence = ex.rb_sequence_2q(
-        ...     target="Q00-Q01",
-        ...     n=100,
-        ...     x90={
-        ...         "Q00": Rect(duration=30, amplitude=0.1),
-        ...         "Q01": Rect(duration=30, amplitude=0.1),
-        ...     },
-        ... )
-
-        >>> sequence = ex.rb_sequence_2q(
-        ...     target="Q00-Q01",
-        ...     n=100,
-        ...     x90={
-        ...         "Q00": Rect(duration=30, amplitude=0.1),
-        ...         "Q01": Rect(duration=30, amplitude=0.1),
-        ...     },
-        ...     interleaved_waveform=Rect(duration=30, amplitude=0.1),
-        ...     interleaved_clifford=Clifford.CNOT(),
-        ... )
-        """
-        target_object = self.experiment_system.get_target(target)
-        if not target_object.is_cr:
-            raise ValueError(f"`{target}` is not a 2Q target.")
-        control_qubit, target_qubit = Target.cr_qubit_pair(target)
-        cr_label = target
-        xi90, ix90 = None, None
-        if isinstance(x90, dict):
-            xi90 = x90.get(control_qubit)
-            ix90 = x90.get(target_qubit)
-        xi90 = xi90 or self.hpi_pulse[control_qubit]
-        ix90 = ix90 or self.hpi_pulse[target_qubit]
-        z90 = VirtualZ(np.pi / 2)
-
-        if zx90 is None:
-            zx90 = self.zx90(
-                control_qubit=control_qubit,
-                target_qubit=target_qubit,
-            )
-
-        if interleaved_waveform is None:
-            cliffords, inverse = self.clifford_generator.create_rb_sequences(
-                n=n,
-                type="2Q",
-                seed=seed,
-            )
-        else:
-            if interleaved_clifford is None:
-                raise ValueError("Interleave map must be provided.")
-            cliffords, inverse = self.clifford_generator.create_irb_sequences(
-                n=n,
-                interleave=interleaved_clifford,
-                type="2Q",
-                seed=seed,
-            )
-
-        with PulseSchedule([control_qubit, cr_label, target_qubit]) as ps:
-
-            def add_gate(gate: str):
-                if gate == "XI90":
-                    ps.add(control_qubit, xi90)
-                elif gate == "IX90":
-                    ps.add(target_qubit, ix90)
-                elif gate == "ZI90":
-                    ps.add(control_qubit, z90)
-                elif gate == "IZ90":
-                    ps.add(target_qubit, z90)
-                    ps.add(cr_label, z90)
-                elif gate == "ZX90":
-                    ps.barrier()
-                    if isinstance(zx90, dict):
-                        ps.add(control_qubit, zx90[control_qubit])
-                        ps.add(target_qubit, zx90[target_qubit])
-                    elif isinstance(zx90, PulseSchedule):
-                        ps.call(zx90)
-                    ps.barrier()
-                else:
-                    raise ValueError("Invalid gate.")
-
-            for clifford in cliffords:
-                for gate in clifford:
-                    add_gate(gate)
-                if interleaved_waveform is not None:
-                    ps.barrier()
-                    if isinstance(interleaved_waveform, dict):
-                        ps.add(control_qubit, interleaved_waveform[control_qubit])
-                        ps.add(target_qubit, interleaved_waveform[target_qubit])
-                    elif isinstance(interleaved_waveform, PulseSchedule):
-                        ps.call(interleaved_waveform)
-                    ps.barrier()
-
-            for gate in inverse:
-                add_gate(gate)
-        return ps
-
-    def rb_experiment_1q(
-        self,
-        *,
-        target: str,
-        n_cliffords_range: ArrayLike | None = None,
-        x90: Waveform | dict[str, Waveform] | None = None,
-        interleaved_waveform: Waveform | None = None,
-        interleaved_clifford: Clifford | dict[str, tuple[complex, str]] | None = None,
-        spectator_state: Literal["0", "1", "+", "-", "+i", "-i"] = "0",
-        seed: int | None = None,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-        save_image: bool = True,
-    ) -> ExperimentResult[RBData]:
-        """
-        Conducts a randomized benchmarking experiment.
-
-        Parameters
-        ----------
-        target : str
-            Target qubit.
-        n_cliffords_range : ArrayLike, optional
-            Range of the number of Cliffords. Defaults to range(0, 1001, 50).
-        x90 : Waveform, optional
-            π/2 pulse used for the experiment. Defaults to None.
-        interleaved_waveform : Waveform, optional
-            Waveform of the interleaved gate. Defaults to None.
-        interleaved_clifford : Clifford | dict[str, tuple[complex, str]], optional
-            Clifford map of the interleaved gate. Defaults to None.
-        spectator_state : Literal["0", "1", "+", "-", "+i", "-i"], optional
-            Spectator state. Defaults to "0".
-        seed : int, optional
-            Random seed.
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-        plot : bool, optional
-            Whether to plot the measured signals. Defaults to True.
-        save_image : bool, optional
-            Whether to save the images. Defaults to True.
-
-        Returns
-        -------
-        ExperimentResult[RBData]
-            Result of the experiment.
-
-        Examples
-        --------
-        >>> result = ex.rb_experiment(
-        ...     target="Q00",
-        ...     n_cliffords_range=range(0, 1001, 50),
-        ...     x90=Rect(duration=30, amplitude=0.1),
-        ... )
-
-        >>> result = ex.rb_experiment(
-        ...     target="Q00",
-        ...     n_cliffords_range=range(0, 1001, 50),
-        ...     x90=Rect(duration=30, amplitude=0.1),
-        ...     interleaved_waveform=Rect(duration=30, amplitude=0.1),
-        ...     interleaved_clifford={
-        ...         "I": (1, "I"),
-        ...         "X": (1, "X"),
-        ...         "Y": (-1, "Y"),
-        ...         "Z": (-1, "Z"),
-        ...     },
-        ... )
-        """
-        if n_cliffords_range is None:
-            n_cliffords_range = np.arange(0, 1001, 50)
-
-        target_object = self.experiment_system.get_target(target)
-        if target_object.is_cr:
-            raise ValueError(f"`{target}` is not a 1Q target.")
-
-        def rb_sequence(N: int) -> PulseSchedule:
-            with PulseSchedule([target]) as ps:
-                # Excite spectator qubits if needed
-                if spectator_state != "0":
-                    spectators = self.get_spectators(target)
-                    for spectator in spectators:
-                        if spectator.label in self.qubit_labels:
-                            pulse = self.get_pulse_for_state(
-                                target=spectator.label,
-                                state=spectator_state,
-                            )
-                            ps.add(spectator.label, pulse)
-                    ps.barrier()
-
-                # Randomized benchmarking sequence
-                rb_sequence = self.rb_sequence_1q(
-                    target=target,
-                    n=N,
-                    x90=x90,
-                    interleaved_waveform=interleaved_waveform,
-                    interleaved_clifford=interleaved_clifford,
-                    seed=seed,
-                )
-                ps.add(
-                    target,
-                    rb_sequence,
-                )
-            return ps
-
-        sweep_result = self.sweep_parameter(
-            rb_sequence,
-            sweep_range=n_cliffords_range,
-            shots=shots,
-            interval=interval,
-            plot=False,
-        )
-
-        sweep_data = sweep_result.data[target]
-
-        fit_result = fitting.fit_rb(
-            target=target,
-            x=sweep_data.sweep_range,
-            y=(sweep_data.normalized + 1) / 2,
-            bounds=((0, 0, 0), (0.5, 1, 1)),
-            title="Randomized benchmarking",
-            xaxis_title="Number of Cliffords",
-            yaxis_title="Normalized signal",
-            xaxis_type="linear",
-            yaxis_type="linear",
-            plot=plot,
-        )
-
-        if save_image:
-            vis.save_figure_image(
-                fit_result["fig"],
-                name=f"rb_{target}",
-            )
-
-        data = {
-            qubit: RBData.new(
-                data,
-                depolarizing_rate=fit_result["depolarizing_rate"],
-                avg_gate_error=fit_result["avg_gate_error"],
-                avg_gate_fidelity=fit_result["avg_gate_fidelity"],
-            )
-            for qubit, data in sweep_result.data.items()
-        }
-
-        return ExperimentResult(data=data)
-
-    def rb_experiment_2q(
-        self,
-        *,
-        target: str,
-        n_cliffords_range: ArrayLike = np.arange(0, 21, 2),
-        x90: TargetMap[Waveform] | None = None,
-        zx90: (
-            PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform] | None
-        ) = None,
-        interleaved_waveform: (
-            PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform] | None
-        ) = None,
-        interleaved_clifford: Clifford | dict[str, tuple[complex, str]] | None = None,
-        spectator_state: Literal["0", "1", "+", "-", "+i", "-i"] = "0",
-        seed: int | None = None,
-        mitigate_readout: bool = True,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-    ):
-        if self.state_centers is None:
-            raise ValueError("State classifiers are not built.")
-
-        n_cliffords_range = np.array(n_cliffords_range, dtype=int)
-
-        target_object = self.experiment_system.get_target(target)
-        if not target_object.is_cr:
-            raise ValueError(f"`{target}` is not a 2Q target.")
-        control_qubit, target_qubit = Target.cr_qubit_pair(target)
-        cr_label = target
-
-        def rb_sequence(N: int) -> PulseSchedule:
-            with PulseSchedule([control_qubit, cr_label, target_qubit]) as ps:
-                # Excite spectator qubits if needed
-                if spectator_state != "0":
-                    control_spectators = {
-                        qubit.label for qubit in self.get_spectators(control_qubit)
-                    }
-                    target_spectators = {
-                        qubit.label for qubit in self.get_spectators(target_qubit)
-                    }
-                    spectators = (control_spectators | target_spectators) - {
-                        control_qubit,
-                        target_qubit,
-                    }
-                    for spectator in spectators:
-                        if spectator in self.qubit_labels:
-                            pulse = self.get_pulse_for_state(
-                                target=spectator,
-                                state=spectator_state,
-                            )
-                            ps.add(spectator, pulse)
-                    ps.barrier()
-
-                # Randomized benchmarking sequence
-                rb_sequence = self.rb_sequence_2q(
-                    target=target,
-                    n=N,
-                    x90=x90,
-                    zx90=zx90,
-                    interleaved_waveform=interleaved_waveform,
-                    interleaved_clifford=interleaved_clifford,
-                    seed=seed,
-                )
-                ps.call(rb_sequence)
-            return ps
-
-        fidelities = []
-
-        for n_clifford in tqdm(n_cliffords_range):
-            result = self.measure(
-                sequence=rb_sequence(n_clifford),
-                mode="single",
-                shots=shots,
-                interval=interval,
-                plot=False,
-            )
-            if mitigate_readout:
-                prob = np.array(list(result.probabilities.values()))
-                cm_inv = self.get_inverse_confusion_matrix(
-                    [control_qubit, target_qubit]
-                )
-                prob_mitigated = prob @ cm_inv
-                p00 = prob_mitigated[0]
-            else:
-                p00 = result.probabilities["00"]
-            fidelities.append(p00)
-
-        fit_result = fitting.fit_rb(
-            target=target,
-            x=n_cliffords_range,
-            y=np.array(fidelities),
-            dimension=4,
-            title="Randomized benchmarking",
-            xaxis_title="Number of Cliffords",
-            yaxis_title="Normalized signal",
-            xaxis_type="linear",
-            yaxis_type="linear",
-            plot=plot,
-        )
-
-        return {
-            "n_cliffords": n_cliffords_range,
-            "fidelities": fidelities,
-            "depolarizing_rate": fit_result["depolarizing_rate"],
-            "avg_gate_error": fit_result["avg_gate_error"],
-            "avg_gate_fidelity": fit_result["avg_gate_fidelity"],
-        }
-
-    def randomized_benchmarking(
-        self,
-        target: str,
-        *,
-        n_cliffords_range: ArrayLike | None = None,
-        n_trials: int = 30,
-        x90: Waveform | dict[str, Waveform] | None = None,
-        zx90: (
-            PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform] | None
-        ) = None,
-        spectator_state: Literal["0", "1", "+", "-", "+i", "-i"] = "0",
-        seeds: ArrayLike | None = None,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-        save_image: bool = True,
-    ) -> dict:
-        """
-        Conducts a randomized benchmarking experiment with multiple trials.
-
-        Parameters
-        ----------
-        target : str
-            Target qubit.
-        n_cliffords_range : ArrayLike, optional
-            Range of the number of Cliffords. Defaults to None.
-        n_trials : int, optional
-            Number of trials for different random seeds. Defaults to 30.
-        x90 : Waveform | dict[str, Waveform], optional
-            π/2 pulse used for the experiment. Defaults to None.
-        zx90 : PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform], optional
-            ZX90 pulses used for 2Q gates. Defaults to None.
-        spectator_state : Literal["0", "1", "+", "-", "+i", "-i"], optional
-            Spectator state. Defaults to "0".
-        seeds : ArrayLike, optional
-            Random seeds. Defaults to None.
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-        plot : bool, optional
-            Whether to plot the measured signals. Defaults to True.
-        save_image : bool, optional
-            Whether to save the images. Defaults to True.
-
-        Returns
-        -------
-        dict
-            Results of the experiment.
-        """
-
-        if seeds is None:
-            seeds = np.random.randint(0, 2**32, n_trials)
-        else:
-            seeds = np.array(seeds, dtype=int)
-            if len(seeds) != n_trials:
-                raise ValueError(
-                    "The number of seeds must be equal to the number of trials."
-                )
-
-        target_object = self.experiment_system.get_target(target)
-        is_2q = target_object.is_cr
-
-        if is_2q:
-            if n_cliffords_range is None:
-                n_cliffords_range = np.arange(0, 21, 2)
-        else:
-            self.validate_rabi_params([target])
-            if n_cliffords_range is None:
-                n_cliffords_range = np.arange(0, 1001, 100)
-
-        n_cliffords_range = np.array(n_cliffords_range, dtype=int)
-
-        results = []
-        for seed in tqdm(seeds):
-            seed = int(seed)
-            with self.util.no_output():
-                if is_2q:
-                    if isinstance(x90, Waveform):
-                        raise ValueError("x90 must be a dict for 2Q gates.")
-                    result = self.rb_experiment_2q(
-                        target=target,
-                        n_cliffords_range=n_cliffords_range,
-                        x90=x90,
-                        zx90=zx90,
-                        interleaved_waveform=None,
-                        interleaved_clifford=None,
-                        spectator_state=spectator_state,
-                        seed=seed,
-                        shots=shots,
-                        interval=interval,
-                        plot=False,
-                    )
-                    signal = result["fidelities"]
-                else:
-                    result = self.rb_experiment_1q(
-                        target=target,
-                        n_cliffords_range=n_cliffords_range,
-                        spectator_state=spectator_state,
-                        x90=x90,
-                        seed=seed,
-                        shots=shots,
-                        interval=interval,
-                        plot=False,
-                        save_image=False,
-                    )
-                    signal = (result.data[target].normalized + 1) / 2
-                results.append(signal)
-
-        mean = np.mean(results, axis=0)
-        std = np.std(results, axis=0)
-
-        fit_result = fitting.fit_rb(
-            target=target,
-            x=n_cliffords_range,
-            y=mean,
-            error_y=std,
-            dimension=4 if is_2q else 2,
-            plot=plot,
-            title="Randomized benchmarking",
-            xaxis_title="Number of Cliffords",
-            yaxis_title="Normalized signal",
-            xaxis_type="linear",
-            yaxis_type="linear",
-        )
-
-        if save_image:
-            vis.save_figure_image(
-                fit_result["fig"],
-                name=f"randomized_benchmarking_{target}",
-            )
-
-        return {
-            "n_cliffords": n_cliffords_range,
-            "mean": mean,
-            "std": std,
-            **fit_result,
-        }
-
-    def interleaved_randomized_benchmarking(
-        self,
-        *,
-        target: str,
-        interleaved_waveform: Waveform | PulseSchedule,
-        interleaved_clifford: Clifford | dict[str, tuple[complex, str]],
-        n_cliffords_range: ArrayLike | None = None,
-        n_trials: int = 30,
-        x90: TargetMap[Waveform] | Waveform | None = None,
-        zx90: (
-            PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform] | None
-        ) = None,
-        spectator_state: Literal["0", "1", "+", "-", "+i", "-i"] = "0",
-        seeds: ArrayLike | None = None,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-        save_image: bool = True,
-    ) -> dict:
-        """
-        Conducts a randomized benchmarking experiment with multiple trials.
-
-        Parameters
-        ----------
-        target : str
-            Target qubit.
-        interleaved_waveform : Waveform
-            Waveform of the interleaved gate.
-        interleaved_clifford : Clifford | dict[str, tuple[complex, str]]
-            Clifford map of the interleaved gate.
-        n_cliffords_range : ArrayLike, optional
-            Range of the number of Cliffords. Defaults to range(0, 1001, 100).
-        n_trials : int, optional
-            Number of trials for different random seeds. Defaults to 30.
-        x90 : Waveform, optional
-            π/2 pulse. Defaults to None.
-        zx90 : PulseSchedule | dict[str, PulseSequence] | dict[str, Waveform], optional
-            ZX90 pulses used for 2Q gates. Defaults to None.
-        spectator_state : Literal["0", "1", "+", "-", "+i", "-i"], optional
-            Spectator state. Defaults to "0".
-        seeds : ArrayLike, optional
-            Random seeds. Defaults to None.
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-        plot : bool, optional
-            Whether to plot the measured signals. Defaults to True.
-        save_image : bool, optional
-            Whether to save the images. Defaults to True.
-
-        Returns
-        -------
-        dict
-            Results of the experiment.
-
-        Examples
-        --------
-        >>> result = ex.interleaved_randomized_benchmarking(
-        ...     target="Q00",
-        ...     interleaved_waveform=Rect(duration=30, amplitude=0.1),
-        ...     interleaved_clifford={
-        ...         "I": (1, "I"),
-        ...         "X": (1, "X"),
-        ...         "Y": (1, "Z"),
-        ...         "Z": (-1, "Y"),
-        ...     },
-        ...     n_cliffords_range=range(0, 1001, 100),
-        ...     n_trials=30,
-        ...     x90=Rect(duration=30, amplitude=0.1),
-        ...     spectator_state="0",
-        ...     show_ref=True,
-        ...     shots=1024,
-        ...     interval=1024,
-        ...     plot=True,
-        ... )
-        """
-        if seeds is None:
-            seeds = np.random.randint(0, 2**32, n_trials)
-        else:
-            seeds = np.array(seeds, dtype=int)
-            if len(seeds) != n_trials:
-                raise ValueError(
-                    "The number of seeds must be equal to the number of trials."
-                )
-
-        target_object = self.experiment_system.get_target(target)
-        is_2q = target_object.is_cr
-
-        if is_2q:
-            if n_cliffords_range is None:
-                n_cliffords_range = np.arange(0, 21, 2)
-        else:
-            self.validate_rabi_params([target])
-            if n_cliffords_range is None:
-                n_cliffords_range = np.arange(0, 1001, 100)
-
-        n_cliffords_range = np.array(n_cliffords_range, dtype=int)
-
-        rb_results = []
-        irb_results = []
-
-        for seed in tqdm(seeds):
-            seed = int(seed)
-            with self.util.no_output():
-                if is_2q:
-                    if isinstance(x90, Waveform):
-                        raise ValueError("x90 must be a dict for 2Q gates.")
-                    if isinstance(zx90, Waveform):
-                        raise ValueError("zx90 must be a dict for 2Q gates.")
-                    if isinstance(interleaved_waveform, Waveform):
-                        raise ValueError(
-                            "interleaved_waveform must be a dict for 2Q gates."
-                        )
-                    rb_result = self.rb_experiment_2q(
-                        target=target,
-                        n_cliffords_range=n_cliffords_range,
-                        x90=x90,
-                        zx90=zx90,
-                        spectator_state=spectator_state,
-                        seed=seed,
-                        shots=shots,
-                        interval=interval,
-                        plot=False,
-                    )
-                    rb_signal = rb_result["fidelities"]
-                    rb_results.append(rb_signal)
-
-                    irb_result = self.rb_experiment_2q(
-                        target=target,
-                        n_cliffords_range=n_cliffords_range,
-                        x90=x90,
-                        zx90=zx90,
-                        interleaved_waveform=interleaved_waveform,
-                        interleaved_clifford=interleaved_clifford,
-                        spectator_state=spectator_state,
-                        seed=seed,
-                        shots=shots,
-                        interval=interval,
-                        plot=False,
-                    )
-                    irb_signal = irb_result["fidelities"]
-                    irb_results.append(irb_signal)
-                else:
-                    rb_result = self.rb_experiment_1q(
-                        target=target,
-                        n_cliffords_range=n_cliffords_range,
-                        x90=x90,  # type: ignore
-                        spectator_state=spectator_state,
-                        seed=seed,
-                        shots=shots,
-                        interval=interval,
-                        plot=False,
-                        save_image=False,
-                    )
-                    rb_signal = (rb_result.data[target].normalized + 1) / 2
-                    rb_results.append(rb_signal)
-
-                    irb_result = self.rb_experiment_1q(
-                        target=target,
-                        n_cliffords_range=n_cliffords_range,
-                        x90=x90,  # type: ignore
-                        interleaved_waveform=interleaved_waveform,  # type: ignore
-                        interleaved_clifford=interleaved_clifford,
-                        spectator_state=spectator_state,
-                        seed=seed,
-                        shots=shots,
-                        interval=interval,
-                        plot=False,
-                        save_image=False,
-                    )
-                    irb_signal = (irb_result.data[target].normalized + 1) / 2
-                    irb_results.append(irb_signal)
-
-        print("Randomized benchmarking:")
-        rb_mean = np.mean(rb_results, axis=0)
-        rb_std = np.std(rb_results, axis=0)
-        rb_fit_result = fitting.fit_rb(
-            target=target,
-            x=n_cliffords_range,
-            y=rb_mean,
-            error_y=rb_std,
-            dimension=4 if is_2q else 2,
-            plot=plot,
-        )
-        A_rb = rb_fit_result["A"]
-        p_rb = rb_fit_result["p"]
-        C_rb = rb_fit_result["C"]
-
-        print("Interleaved randomized benchmarking:")
-        irb_mean = np.mean(irb_results, axis=0)
-        irb_std = np.std(irb_results, axis=0)
-        irb_fit_result = fitting.fit_rb(
-            target=target,
-            x=n_cliffords_range,
-            y=irb_mean,
-            error_y=irb_std,
-            dimension=4 if is_2q else 2,
-            plot=plot,
-            title="Interleaved randomized benchmarking",
-        )
-        A_irb = irb_fit_result["A"]
-        p_irb = irb_fit_result["p"]
-        C_irb = irb_fit_result["C"]
-
-        dimension = 4 if is_2q else 2
-        gate_error = (dimension - 1) * (1 - (p_irb / p_rb)) / dimension
-        gate_fidelity = 1 - gate_error
-
-        fig = fitting.plot_irb(
-            target=target,
-            x=n_cliffords_range,
-            y_rb=rb_mean,
-            y_irb=irb_mean,
-            error_y_rb=rb_std,
-            error_y_irb=irb_std,
-            A_rb=A_rb,
-            A_irb=A_irb,
-            p_rb=p_rb,
-            p_irb=p_irb,
-            C_rb=C_rb,
-            C_irb=C_irb,
-            gate_fidelity=gate_fidelity,
-            title="Interleaved randomized benchmarking",
-            xaxis_title="Number of Cliffords",
-            yaxis_title="Normalized signal",
-        )
-        if save_image:
-            vis.save_figure_image(
-                fig,
-                name=f"interleaved_randomized_benchmarking_{target}",
-            )
-
-        print("")
-        print(f"Gate error: {gate_error * 100:.3f}%")
-        print(f"Gate fidelity: {gate_fidelity * 100:.3f}%")
-        print("")
-
-        return {
-            "gate_error": gate_error,
-            "gate_fidelity": gate_fidelity,
-            "rb_fit_result": rb_fit_result,
-            "irb_fit_result": irb_fit_result,
-            "fig": fig,
-        }
-
-    def state_tomography(
-        self,
-        sequence: TargetMap[IQArray] | TargetMap[Waveform] | PulseSchedule,
-        *,
-        x90: TargetMap[Waveform] | None = None,
-        initial_state: TargetMap[str] | None = None,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = False,
-    ) -> dict[str, tuple[float, float, float]]:
-        """
-        Conducts a state tomography experiment.
-
-        Parameters
-        ----------
-        sequence : TargetMap[IQArray] | TargetMap[Waveform] | PulseSchedule
-            Sequence to measure for each target.
-        x90 : TargetMap[Waveform], optional
-            π/2 pulse. Defaults to None.
-        initial_state : TargetMap[str], optional
-            Initial state of each target. Defaults to None.
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-        plot : bool, optional
-            Whether to plot the measured signals. Defaults to False.
-
-        Returns
-        -------
-        dict[str, tuple[float, float, float]]
-            Results of the experiment.
-        """
-        if isinstance(sequence, PulseSchedule):
-            sequence = sequence.get_sequences()
-        else:
-            sequence = {
-                target: (
-                    Pulse(waveform) if not isinstance(waveform, Waveform) else waveform
-                )
-                for target, waveform in sequence.items()
-            }
-
-        x90 = x90 or self.hpi_pulse
-
-        buffer: dict[str, list[float]] = defaultdict(list)
-
-        qubits = set(Target.qubit_label(target) for target in sequence)
-        targets = list(qubits | sequence.keys())
-
-        for basis in ["X", "Y", "Z"]:
-            with PulseSchedule(targets) as ps:
-                # Initialization pulses
-                if initial_state is not None:
-                    for qubit in qubits:
-                        if qubit in initial_state:
-                            init_pulse = self.get_pulse_for_state(
-                                target=qubit,
-                                state=initial_state[qubit],
-                            )
-                            ps.add(qubit, init_pulse)
-                    ps.barrier()
-
-                # Pulse sequences
-                for target, waveform in sequence.items():
-                    ps.add(target, waveform)
-                ps.barrier()
-
-                # Basis transformation pulses
-                for qubit in qubits:
-                    x90p = x90[qubit]
-                    y90m = x90p.shifted(-np.pi / 2)
-                    if basis == "X":
-                        ps.add(qubit, y90m)
-                    elif basis == "Y":
-                        ps.add(qubit, x90p)
-
-            measure_result = self.measure(
-                ps,
-                shots=shots,
-                interval=interval,
-                plot=plot,
-            )
-            for qubit, data in measure_result.data.items():
-                rabi_param = self.rabi_params[qubit]
-                if rabi_param is None:
-                    raise ValueError("Rabi parameters are not stored.")
-                values = data.kerneled
-                values_rotated = values * np.exp(-1j * rabi_param.angle)
-                values_normalized = (
-                    np.imag(values_rotated) - rabi_param.offset
-                ) / rabi_param.amplitude
-                buffer[qubit] += [values_normalized]
-
-        result = {
-            qubit: (
-                values[0],  # X
-                values[1],  # Y
-                values[2],  # Z
-            )
-            for qubit, values in buffer.items()
-        }
-        return result
-
-    def state_evolution_tomography(
-        self,
-        *,
-        sequences: (
-            Sequence[TargetMap[IQArray]]
-            | Sequence[TargetMap[Waveform]]
-            | Sequence[PulseSchedule]
-        ),
-        x90: TargetMap[Waveform] | None = None,
-        initial_state: TargetMap[str] | None = None,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-    ) -> dict[str, NDArray[np.float64]]:
-        """
-        Conducts a state evolution tomography experiment.
-
-        Parameters
-        ----------
-        sequences : Sequence[TargetMap[IQArray]] | Sequence[TargetMap[Waveform]] | Sequence[PulseSchedule]
-            Sequences to measure for each target.
-        x90 : TargetMap[Waveform], optional
-            π/2 pulse. Defaults to None.
-        initial_state : TargetMap[str], optional
-            Initial state of each target. Defaults to None.
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-        plot : bool, optional
-            Whether to plot the measured signals. Defaults to False.
-
-        Returns
-        -------
-        dict[str, NDArray[np.float64]]
-            Results of the experiment.
-        """
-        buffer: dict[str, list[tuple[float, float, float]]] = defaultdict(list)
-        for sequence in tqdm(sequences):
-            state_vectors = self.state_tomography(
-                sequence=sequence,
-                x90=x90,
-                initial_state=initial_state,
-                shots=shots,
-                interval=interval,
-                plot=False,
-            )
-            for target, state_vector in state_vectors.items():
-                buffer[target].append(state_vector)
-
-        result = {target: np.array(states) for target, states in buffer.items()}
-
-        if plot:
-            for target, states in result.items():
-                print(f"State evolution : {target}")
-                vis.display_bloch_sphere(states)
-
-        return result
-
-    def pulse_tomography(
-        self,
-        waveforms: TargetMap[IQArray] | TargetMap[Waveform] | PulseSchedule,
-        *,
-        x90: TargetMap[Waveform] | None = None,
-        initial_state: TargetMap[str] | None = None,
-        n_samples: int = 100,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-    ) -> TargetMap[NDArray[np.float64]]:
-        """
-        Conducts a pulse tomography experiment.
-
-        Parameters
-        ----------
-        waveforms : TargetMap[IQArray] | TargetMap[Waveform] | PulseSchedule
-            Waveforms to measure for each target.
-        x90 : TargetMap[Waveform], optional
-            π/2 pulse. Defaults to None.
-        initial_state : TargetMap[str], optional
-            Initial state of each target. Defaults to None.
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-        plot : bool, optional
-            Whether to plot the measured signals. Defaults to True.
-        """
-        self.validate_rabi_params()
-
-        if isinstance(waveforms, PulseSchedule):
-            sequences = waveforms.get_sequences()
-        else:
-            sequences = waveforms
-
-        pulses: dict[str, Waveform] = {}
-        pulse_length_set = set()
-        for target, waveform in sequences.items():
-            if isinstance(waveform, Waveform):
-                pulse = waveform
-            elif isinstance(waveform, list) or isinstance(waveform, np.ndarray):
-                pulse = Pulse(waveform)
-            else:
-                raise ValueError("Invalid waveform.")
-            pulses[target] = pulse
-            pulse_length_set.add(pulse.length)
-        if len(pulse_length_set) != 1:
-            raise ValueError("The lengths of the waveforms must be the same.")
-
-        pulse_length = pulse_length_set.pop()
-
-        if plot:
-            if isinstance(waveforms, PulseSchedule):
-                waveforms.plot(title="Pulse sequence")
-            else:
-                for target in pulses:
-                    pulses[target].plot(title=f"Waveform : {target}")
-
-        def partial_waveform(waveform: Waveform, index: int) -> Waveform:
-            """Returns a partial waveform up to the given index."""
-
-            # If the waveform is a PulseSequence, we need to handle the PhaseShift gate.
-            if isinstance(waveform, PulseSequence):
-                current_index = 0
-                pulse_sequence = PulseSequence([])
-                for pulse in waveform._sequence:
-                    # If the pulse is a PhaseShift gate, we can simply add it to the sequence.
-                    if isinstance(pulse, PhaseShift):
-                        pulse_sequence = pulse_sequence.added(pulse)
-                        continue
-                    # If the pulse is a Pulse and the length is greater than the index, we need to create a partial pulse.
-                    elif current_index + pulse.length > index:
-                        pulse = Pulse(pulse.values[0 : index - current_index])
-                        pulse_sequence = pulse_sequence.added(pulse)
-                        break
-                    # If the pulse is a Pulse and the length is less than the index, we can add the pulse to the sequence.
-                    else:
-                        pulse_sequence = pulse_sequence.added(pulse)
-                        current_index += pulse.length
-                return pulse_sequence
-            # If the waveform is a Pulse, we can simply return the partial waveform.
-            else:
-                return Pulse(waveform.values[0:index])
-
-        if n_samples < pulse_length:
-            indices = np.linspace(0, pulse_length - 1, n_samples).astype(int)
-        else:
-            indices = np.arange(pulse_length)
-
-        sequences = [
-            {target: partial_waveform(pulse, i) for target, pulse in pulses.items()}
-            for i in indices
-        ]
-
-        result = self.state_evolution_tomography(
-            sequences=sequences,
-            x90=x90,
-            initial_state=initial_state,
-            shots=shots,
-            interval=interval,
-            plot=plot,
-        )
-
-        if plot:
-            times = pulses.popitem()[1].times[indices]
-            for target, states in result.items():
-                vis.plot_bloch_vectors(
-                    times=times,
-                    bloch_vectors=states,
-                    title=f"State evolution : {target}",
-                )
-
-        return result
-
-    def measure_population(
-        self,
-        sequence: TargetMap[IQArray] | TargetMap[Waveform] | PulseSchedule,
-        *,
-        fit_gmm: bool = False,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-    ) -> tuple[dict[str, NDArray[np.float64]], dict[str, NDArray[np.float64]]]:
-        """
-        Measures the state populations of the target qubits.
-
-        Parameters
-        ----------
-        sequence : TargetMap[IQArray] | TargetMap[Waveform] | PulseSchedule
-            Sequence to measure for each target.
-        fit_gmm : bool, optional
-            Whether to fit the data with a Gaussian mixture model. Defaults to False
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        tuple[dict[str, NDArray[np.float64]], dict[str, NDArray[np.float64]]]
-            State probabilities and standard deviations.
-
-        Examples
-        --------
-        >>> sequence = {
-        ...     "Q00": ex.hpi_pulse["Q00"],
-        ...     "Q01": ex.hpi_pulse["Q01"],
-        ... }
-        >>> result = ex.measure_population(sequence)
-        """
-        if self.classifiers is None:
-            raise ValueError("Classifiers are not built. Run `build_classifier` first.")
-
-        result = self.measure(
-            sequence,
-            mode="single",
-            shots=shots,
-            interval=interval,
-        )
-        if fit_gmm:
-            probabilities = {
-                target: self.classifiers[target].estimate_weights(
-                    result.data[target].kerneled
-                )
-                for target in result.data
-            }
-            standard_deviations = {
-                target: np.zeros_like(probabilities[target]) for target in probabilities
-            }
-        else:
-            probabilities = {
-                target: data.probabilities for target, data in result.data.items()
-            }
-            standard_deviations = {
-                target: data.standard_deviations for target, data in result.data.items()
-            }
-        return probabilities, standard_deviations
-
-    def measure_population_dynamics(
-        self,
-        *,
-        sequence: ParametricPulseSchedule | ParametricWaveformDict,
-        params_list: Sequence | NDArray,
-        fit_gmm: bool = False,
-        xlabel: str = "Index",
-        scatter_mode: str = "lines+markers",
-        show_error: bool = True,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-    ) -> tuple[dict[str, NDArray[np.float64]], dict[str, NDArray[np.float64]]]:
-        """
-        Measures the population dynamics of the target qubits.
-
-        Parameters
-        ----------
-        sequence : ParametricPulseSchedule | ParametricWaveformDict
-            Parametric sequence to measure.
-        params_list : Sequence | NDArray
-            List of parameters.
-        fit_gmm : bool, optional
-            Whether to fit the data with a Gaussian mixture model. Defaults to False.
-        xlabel : str, optional
-            Label of the x-axis. Defaults to "Index".
-        shots : int, optional
-            Number of shots. Defaults to DEFAULT_SHOTS.
-        interval : int, optional
-            Interval between shots. Defaults to DEFAULT_INTERVAL.
-
-        Returns
-        -------
-        tuple[dict[str, NDArray[np.float64]], dict[str, NDArray[np.float64]]]
-            State probabilities and standard deviations.
-
-        Examples
-        --------
-        >>> sequence = lambda x: {
-        ...     "Q00": ex.hpi_pulse["Q00"].scaled(x),
-        ...     "Q01": ex.hpi_pulse["Q01"].scaled(x),
-        >>> params_list = np.linspace(0.5, 1.5, 100)
-        >>> result = ex.measure_popultion_dynamics(sequence, params_list)
-        """
-        if isinstance(params_list[0], int):
-            x = params_list
-        else:
-            try:
-                float(params_list[0])
-                x = params_list
-            except ValueError:
-                x = np.arange(len(params_list))
-
-        buffer_pops = defaultdict(list)
-        buffer_errs = defaultdict(list)
-
-        for params in tqdm(params_list):
-            prob_dict, err_dict = self.measure_population(
-                sequence=sequence(params),
-                fit_gmm=fit_gmm,
-                shots=shots,
-                interval=interval,
-            )
-            for target, probs in prob_dict.items():
-                buffer_pops[target].append(probs)
-            for target, errors in err_dict.items():
-                buffer_errs[target].append(errors)
-
-        result_pops = {
-            target: np.array(buffer_pops[target]).T for target in buffer_pops
-        }
-        result_errs = {
-            target: np.array(buffer_errs[target]).T for target in buffer_errs
-        }
-
-        for target in result_pops:
-            fig = go.Figure()
-            for state, probs in enumerate(result_pops[target]):
-                fig.add_scatter(
-                    name=f"|{state}⟩",
-                    mode=scatter_mode,
-                    x=x,
-                    y=probs,
-                    error_y=(
-                        dict(
-                            type="data",
-                            array=result_errs[target][state],
-                            visible=True,
-                            thickness=1.5,
-                            width=3,
-                        )
-                        if show_error
-                        else None
-                    ),
-                    marker=dict(size=5),
-                )
-            fig.update_layout(
-                title=f"Population dynamics : {target}",
-                xaxis_title=xlabel,
-                yaxis_title="Probability",
-                yaxis_range=[-0.1, 1.1],
-            )
-            fig.show()
-
-        return result_pops, result_errs
 
     def measure_cr_dynamics(
         self,
@@ -3092,146 +1343,6 @@ class CalibrationMixin(
             "duration_range": duration_range,
             "signal": signal,
             **fit_result,
-        }
-
-    def zx90(
-        self,
-        control_qubit: str,
-        target_qubit: str,
-        cr_duration: float | None = None,
-        cr_ramptime: float | None = None,
-        cr_amplitude: float | None = None,
-        cr_phase: float | None = None,
-        cancel_amplitude: float | None = None,
-        cancel_phase: float | None = None,
-        echo: bool = True,
-        x180: TargetMap[Waveform] | Waveform | None = None,
-    ) -> PulseSchedule:
-        cr_label = f"{control_qubit}-{target_qubit}"
-        cr_params = self.system_note.get(CR_PARAMS)[cr_label]
-
-        if x180 is None:
-            if control_qubit in self.drag_pi_pulse:
-                pi_pulse = self.drag_pi_pulse[control_qubit]
-            else:
-                pi_pulse = self.pi_pulse[control_qubit]
-        elif isinstance(x180, Waveform):
-            pi_pulse = x180
-        else:
-            pi_pulse = x180[control_qubit]
-
-        if cr_amplitude is not None and cancel_amplitude is None:
-            cr_cancel_ratio = (
-                cr_params["cancel_pulse"]["amplitude"]
-                / cr_params["cr_pulse"]["amplitude"]
-            )
-            cancel_amplitude = cr_amplitude * cr_cancel_ratio
-
-        return CrossResonance(
-            control_qubit=control_qubit,
-            target_qubit=target_qubit,
-            cr_amplitude=cr_amplitude or cr_params["cr_pulse"]["amplitude"],
-            cr_duration=cr_duration or cr_params["duration"],
-            cr_ramptime=cr_ramptime or cr_params["ramptime"],
-            cr_phase=cr_phase or cr_params["cr_pulse"]["phase"],
-            cancel_amplitude=cancel_amplitude or cr_params["cancel_pulse"]["amplitude"],
-            cancel_phase=cancel_phase or cr_params["cancel_pulse"]["phase"],
-            echo=echo,
-            pi_pulse=pi_pulse,
-        )
-
-    def cnot(
-        self,
-        control_qubit: str,
-        target_qubit: str,
-        zx90: PulseSchedule | None = None,
-        x90: Waveform | None = None,
-    ) -> PulseSchedule:
-        cr_label = f"{control_qubit}-{target_qubit}"
-        zx90 = zx90 or self.zx90(control_qubit, target_qubit)
-        if x90 is None:
-            if target_qubit in self.drag_hpi_pulse:
-                x90 = self.drag_hpi_pulse[target_qubit]
-            else:
-                x90 = self.hpi_pulse[target_qubit]
-
-        with PulseSchedule([control_qubit, cr_label, target_qubit]) as ps:
-            ps.call(zx90)
-            ps.add(control_qubit, VirtualZ(-np.pi / 2))
-            ps.add(target_qubit, x90.scaled(-1))
-
-        return ps
-
-    def measure_bell_state(
-        self,
-        control_qubit: str,
-        target_qubit: str,
-        zx90: PulseSchedule | None = None,
-        shots: int = DEFAULT_SHOTS,
-        interval: int = DEFAULT_INTERVAL,
-        plot: bool = True,
-        save_image: bool = False,
-    ) -> dict:
-        if self.state_centers is None:
-            self.build_classifier(plot=False)
-
-        cnot = self.cnot(
-            control_qubit=control_qubit,
-            target_qubit=target_qubit,
-            zx90=zx90,
-        )
-        result = self.measure(
-            cnot,
-            initial_states={control_qubit: "+"},
-            mode="single",
-            shots=shots,
-            interval=interval,
-        )
-
-        labels = [f"|{i}⟩" for i in result.probabilities.keys()]
-        prob = np.array(list(result.probabilities.values()))
-        cm_imv = self.get_inverse_confusion_matrix([control_qubit, target_qubit])
-
-        mitigated_prob = prob @ cm_imv
-
-        fig = go.Figure()
-        fig.add_trace(
-            go.Bar(
-                x=labels,
-                y=prob,
-                name="Raw",
-            )
-        )
-        fig.add_trace(
-            go.Bar(
-                x=labels,
-                y=mitigated_prob,
-                name="Mitigated",
-            )
-        )
-        fig.update_layout(
-            title=f"Bell state measurement: {control_qubit}-{target_qubit}",
-            xaxis_title="State label",
-            yaxis_title="Probability",
-            barmode="group",
-            yaxis_range=[0, 1],
-        )
-        if plot:
-            fig.show()
-
-        if save_image:
-            vis.save_figure_image(
-                fig,
-                f"bell_state_measurement_{control_qubit}-{target_qubit}",
-            )
-
-        for label, p, mp in zip(labels, prob, mitigated_prob):
-            print(f"{label} : {p:.2%} -> {mp:.2%}")
-
-        return {
-            "raw": prob,
-            "mitigated": mitigated_prob,
-            "figure": fig,
         }
 
     def optimize_x90(
